migrate: upgrade from chaindb 4 to 6
